Commit8

+++ java-school-fullstack/src/main/java/mthree/com/fullstackschool/service/StudentServiceImpl.java
@@ -38,7 +38,14 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student getStudentById(int id) {
         //YOUR CODE STARTS HERE
 
-        return studentDao.findStudentById(id);
+        try {
+            return studentDao.findStudentById(id);
+        } catch(DataAccessException ex) {
+            Student student = new Student();
+            student.setStudentFirstName("Student Not Found");
+            student.setStudentLastName("Student Not Found");
+            return student;
+        }
 
         //YOUR CODE ENDS HERE
     }
@@ -46,6 +53,11 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student addNewStudent(Student student) {
         //YOUR CODE STARTS HERE
 
+        if(student.getStudentFirstName() == null || student.getStudentLastName() == null) {
+            student.setStudentFirstName("First Name blank, student NOT added");
+            student.setStudentLastName("Last Name blank, student NOT added");
+            return student;
+        }
         return studentDao.createNewStudent(student);
 
         //YOUR CODE ENDS HERE
@@ -54,9 +66,11 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public Student updateStudentData(int id, Student student) {
         //YOUR CODE STARTS HERE
 
-        student.setStudentId(id);
+        if(id != student.getStudentId()) {
+            student.setStudentFirstName("IDs do not match, student not updated");
+            student.setStudentLastName("IDs do not match, student not updated");
+        }
         studentDao.updateStudent(student);
-
         return student;
 
         //YOUR CODE ENDS HERE
@@ -73,7 +87,17 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void deleteStudentFromCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        studentDao.deleteStudentFromCourse(studentId, courseId);
+        Student student = studentDao.findStudentById(studentId);
+        Course course = courseDao.findCourseById(courseId);
+
+        if(student.getStudentFirstName().equals("Student Not Found")) {
+            System.out.println("Student not found");
+        } else if(course.getCourseName().equals("Course Not Found")) {
+            System.out.println("Course not found");
+        } else {
+            studentDao.deleteStudentFromCourse(studentId, courseId);
+            System.out.println("Student: " + studentId + " deleted from course: " + courseId);
+        }
 
         //YOUR CODE ENDS HERE
     }
@@ -81,7 +105,21 @@ public class StudentServiceImpl implements StudentServiceInterface {
     public void addStudentToCourse(int studentId, int courseId) {
         //YOUR CODE STARTS HERE
 
-        studentDao.addStudentToCourse(studentId, courseId);
+        Student student = studentDao.findStudentById(studentId);
+        Course course = courseDao.findCourseById(courseId);
+
+        if(student.getStudentFirstName().equals("Student Not Found")) {
+            System.out.println("Student not found");
+        } else if(course.getCourseName().equals("Course Not Found")) {
+            System.out.println("Course not found");
+        } else {
+            try {
+                studentDao.addStudentToCourse(studentId, courseId);
+                System.out.println("Student: " + studentId + " added to course: " + courseId);
+            } catch (DataAccessException ex) {
+                System.out.println("Student: " + studentId + " already enrolled in course: " + courseId);
+            }
+        }
 
         //YOUR CODE ENDS HERE
     }
